pyproject.toml
[project]
name = "computer2mqtt"
version = "0.1.4"
description = "This project is a multi-system and multi-OS adaptation of mac2mqtt. It allows you to execute any command on your system, providing endless possibilities for automation and control."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomqtt>=2.3.0",
    "click>=8.1.7",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.3",
    "pre-commit>=3.3.3",
    "pytest",
    "pytest-asyncio<0.23",
    "pytest-cov",
]

release = [
    "twine",
]

[tool.pyright]
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.isort]
profile = "black"
src_paths = "tests"
line_length = 80
include_trailing_comma = true
multi_line_output = 3

[tool.ruff]
select = ["E", "F", "B"]
exclude = ["__pycache__", "terraform", ".git", ".venv"]
ignore = ["E501"]

[project.scripts]
computer2mqtt = "computer2mqtt:sync_main"


[build-system]
requires = ["hatchling", "twine"]
build-backend = "hatchling.build"
